backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation & Settings
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# CORS
starlette==0.41.3

# Utils
python-dotenv==1.0.1
aiofiles==24.1.0
uuid6==2024.7.10
